forumbackend/requirements.txt
Django==3.2.19
django-cors-headers==4.0.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
Pillow==9.5.0

